@@ -18,9 +18,7 @@ def compileAvroVersion = "1.9.0"
 dependencies {
     compile localGroovy()
     compile "org.apache.avro:avro-compiler:${compileAvroVersion}"
-    testCompile("org.spockframework:spock-core:1.2-groovy-2.4") {
-        exclude module: "groovy-all"
-    }
+    testCompile "org.spockframework:spock-core:1.3-groovy-2.5"
     testCompile gradleTestKit()
 }
 
@@ -30,10 +28,14 @@ tasks.withType(AbstractCompile) {
 tasks.withType(JavaCompile) {
     options.compilerArgs << "-Xlint:all" << "-Xlint:-options" << "-Werror"
 }
-sourceCompatibility = 1.7
 
 version = "0.17.1-SNAPSHOT"
 group = "com.commercehub.gradle.plugin"
+
+tasks.withType(AbstractArchiveTask) {
+    preserveFileTimestamps = false
+    reproducibleFileOrder = true
+}
 
 task sourcesJar(type: Jar, dependsOn: classes) {
     from sourceSets.main.allSource
@@ -137,15 +139,11 @@ dependencies {
 }
 
 checkstyle {
-    toolVersion = "6.1.1" // Last version of checkstyle to support Java 6
-}
-// Nasty workaround for https://issues.gradle.org/browse/GRADLE-2888
-def checkstyleWarningsFile = "build/reports/checkstyle/main.xml"
-checkstyleMain.doLast {
-    File warningsFile = file(checkstyleWarningsFile)
-    if (warningsFile.exists() && warningsFile.text.contains("<error ")) {
-        throw new GradleException("There were checkstyle warnings! For more info check ${warningsFile}")
-    }
+    ignoreFailures = false
+    maxErrors = 0
+    maxWarnings = 0
+    showViolations = true
+    toolVersion = "8.23"
 }
 // In Gradle 4.8 the checkstyle basedir changed to no longer be the project root by default; thus we need to specify
 checkstyleMain {
@@ -153,8 +151,12 @@ checkstyleMain {
 }
 
 codenarc {
-    toolVersion = "1.0"
     config = project.resources.text.fromFile("config/codenarc/codenarc.groovy")
+    ignoreFailures = false
+    maxPriority1Violations = 0
+    maxPriority2Violations = 0
+    maxPriority3Violations = 0
+    toolVersion = "1.4"
 }
 
 tasks.create("testVersionCompatibility") {
@@ -174,16 +176,18 @@ test {
     ]
 }
 
+// Java 8+ is required due to Avro 1.9.0 requirements
+// Java 8+ is also required by Gradle 5.x
+sourceCompatibility = 8
+
 def avroVersions = ["1.9.0"]
 def gradleVersions = []
-if (!org.gradle.api.JavaVersion.current().isJava9Compatible()) { // Gradle 4.2.1 is the first version that supports modern Java 9, as per https://discuss.gradle.org/t/could-not-determine-java-version-from-9-0-1/24457
-    gradleVersions.addAll("3.0", "3.1", "3.2", "3.2.1", "3.3", "3.4", "3.4.1", "3.5", "3.5.1", "4.0", "4.0.1", "4.0.2", "4.1", "4.2")
-}
 if (!org.gradle.api.JavaVersion.current().isJava11Compatible()) { // Gradle 4.8 appears to be the first version that supports Java 11, as per https://github.com/gradle/gradle/pull/4759 and testing
-    gradleVersions.addAll("4.2.1", "4.3", "4.3.1", "4.4", "4.4.1", "4.5", "4.5.1", "4.6", "4.7")
+    gradleVersions.addAll("4.4", "4.4.1", "4.5", "4.5.1", "4.6", "4.7")
 }
-gradleVersions.addAll("4.8", "4.8.1", "4.9", "4.10", "4.10.1", "4.10.2")
-// Support for running Gradle on Java 7 is deprecated and will be removed in Gradle 5, as per https://docs.gradle.org/4.2.1/release-notes.html
+gradleVersions.addAll("4.8", "4.8.1", "4.9", "4.10", "4.10.1", "4.10.2", "4.10.3")
+gradleVersions.addAll("5.0", "5.1", "5.1.1", "5.2", "5.2.1", "5.3", "5.3.1", "5.4", "5.4.1", "5.5", "5.5.1", "5.6")
+
 
 avroVersions.each { def avroVersion ->
     gradleVersions.each { def gradleVersion ->
