@@ -1,5 +1,5 @@
 /**
- * Copyright © 2013-2015 Commerce Technologies, LLC.
+ * Copyright © 2013-2019 Commerce Technologies, LLC.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,16 +15,14 @@
  */
 package com.commercehub.gradle.plugin.avro;
 
-import org.gradle.api.Action;
+import java.io.File;
+import java.io.FileFilter;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.gradle.api.Task;
-import org.gradle.api.internal.ConventionMapping;
-import org.gradle.api.internal.IConventionAware;
-import org.gradle.api.plugins.AppliedPlugin;
+import org.gradle.api.file.Directory;
 import org.gradle.api.plugins.JavaPlugin;
 import org.gradle.api.plugins.JavaPluginConvention;
-import org.gradle.api.specs.Spec;
+import org.gradle.api.provider.Provider;
 import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.SourceSetContainer;
 import org.gradle.api.tasks.SourceTask;
@@ -33,11 +31,12 @@ import org.gradle.plugins.ide.idea.GenerateIdeaModule;
 import org.gradle.plugins.ide.idea.IdeaPlugin;
 import org.gradle.plugins.ide.idea.model.IdeaModule;
 
-import java.io.File;
-import java.io.FileFilter;
-import java.util.concurrent.Callable;
-
-import static com.commercehub.gradle.plugin.avro.Constants.*;
+import static com.commercehub.gradle.plugin.avro.Constants.GROUP_SOURCE_GENERATION;
+import static com.commercehub.gradle.plugin.avro.Constants.IDL_EXTENSION;
+import static com.commercehub.gradle.plugin.avro.Constants.JAVA_EXTENSION;
+import static com.commercehub.gradle.plugin.avro.Constants.PROTOCOL_EXTENSION;
+import static com.commercehub.gradle.plugin.avro.Constants.SCHEMA_EXTENSION;
+import static com.commercehub.gradle.plugin.avro.GradleCompatibility.configurePropertyConvention;
 
 public class AvroPlugin implements Plugin<Project> {
     @Override
@@ -49,56 +48,38 @@ public class AvroPlugin implements Plugin<Project> {
     }
 
     private static void configureTasks(final Project project) {
-        getSourceSets(project).all(new Action<SourceSet>() {
-            public void execute(SourceSet sourceSet) {
-                GenerateAvroProtocolTask protoTask = configureProtocolGenerationTask(project, sourceSet);
-                GenerateAvroJavaTask javaTask = configureJavaGenerationTask(project, sourceSet, protoTask);
-                configureTaskDependencies(project, sourceSet, javaTask);
-            }
+        getSourceSets(project).all(sourceSet -> {
+            GenerateAvroProtocolTask protoTask = configureProtocolGenerationTask(project, sourceSet);
+            GenerateAvroJavaTask javaTask = configureJavaGenerationTask(project, sourceSet, protoTask);
+            configureTaskDependencies(project, sourceSet, javaTask);
         });
     }
 
     private static void configureIntelliJ(final Project project) {
-        project.getPlugins().withType(IdeaPlugin.class).all(new Action<IdeaPlugin>() {
-            @Override
-            public void execute(IdeaPlugin ideaPlugin) {
-                SourceSet mainSourceSet = getMainSourceSet(project);
-                SourceSet testSourceSet = getTestSourceSet(project);
-                project.getTasks().withType(GenerateIdeaModule.class).all(new Action<GenerateIdeaModule>() {
-                    @Override
-                    public void execute(final GenerateIdeaModule generateIdeaModule) {
-                        project.getTasks().withType(GenerateAvroJavaTask.class).all(new Action<GenerateAvroJavaTask>() {
-                            @Override
-                            public void execute(final GenerateAvroJavaTask generateAvroJavaTask) {
-                                generateIdeaModule.doFirst(new Action<Task>() {
-                                    @Override
-                                    public void execute(Task task) {
-                                        project.mkdir(generateAvroJavaTask.getOutputDir());
-                                    }
-                                });
-                            }
-                        });
-                    }
-                });
-                IdeaModule module = ideaPlugin.getModel().getModule();
-                module.setSourceDirs(new SetBuilder<File>()
-                    .addAll(module.getSourceDirs())
-                    .add(getAvroSourceDir(project, mainSourceSet))
-                    .build());
-                module.setTestSourceDirs(new SetBuilder<File>()
-                    .addAll(module.getTestSourceDirs())
-                    .add(getAvroSourceDir(project, testSourceSet))
-                    .build());
-                // IntelliJ doesn't allow source directories beneath an excluded directory.
-                // Thus, we remove the build directory exclude and add all non-generated sub-directories as excludes.
-                SetBuilder<File> excludeDirs = new SetBuilder<>();
-                excludeDirs.addAll(module.getExcludeDirs()).remove(project.getBuildDir());
-                File buildDir = project.getBuildDir();
-                if (buildDir.isDirectory()) {
-                    excludeDirs.addAll(project.getBuildDir().listFiles(new NonGeneratedDirectoryFileFilter()));
-                }
-                module.setExcludeDirs(excludeDirs.build());
+        project.getPlugins().withType(IdeaPlugin.class).all(ideaPlugin -> {
+            SourceSet mainSourceSet = getMainSourceSet(project);
+            SourceSet testSourceSet = getTestSourceSet(project);
+            project.getTasks().withType(GenerateIdeaModule.class).all(generateIdeaModule ->
+                project.getTasks().withType(GenerateAvroJavaTask.class).all(generateAvroJavaTask ->
+                    generateIdeaModule.doFirst(task -> project.mkdir(generateAvroJavaTask.getOutputDir()))));
+            IdeaModule module = ideaPlugin.getModel().getModule();
+            module.setSourceDirs(new SetBuilder<File>()
+                .addAll(module.getSourceDirs())
+                .add(getAvroSourceDir(project, mainSourceSet))
+                .build());
+            module.setTestSourceDirs(new SetBuilder<File>()
+                .addAll(module.getTestSourceDirs())
+                .add(getAvroSourceDir(project, testSourceSet))
+                .build());
+            // IntelliJ doesn't allow source directories beneath an excluded directory.
+            // Thus, we remove the build directory exclude and add all non-generated sub-directories as excludes.
+            SetBuilder<File> excludeDirs = new SetBuilder<>();
+            excludeDirs.addAll(module.getExcludeDirs()).remove(project.getBuildDir());
+            File buildDir = project.getBuildDir();
+            if (buildDir.isDirectory()) {
+                excludeDirs.addAll(project.getBuildDir().listFiles(new NonGeneratedDirectoryFileFilter()));
             }
+            module.setExcludeDirs(excludeDirs.build());
         });
     }
 
@@ -110,12 +91,7 @@ public class AvroPlugin implements Plugin<Project> {
         task.setGroup(GROUP_SOURCE_GENERATION);
         task.source(getAvroSourceDir(project, sourceSet));
         task.include("**/*." + IDL_EXTENSION);
-        task.getConventionMapping().map("outputDir", new Callable<File>() {
-            @Override
-            public File call() throws Exception {
-                return getGeneratedOutputDir(project, sourceSet, PROTOCOL_EXTENSION);
-            }
-        });
+        configurePropertyConvention(task.getOutputDir(), getGeneratedOutputDir(project, sourceSet, PROTOCOL_EXTENSION));
         return task;
     }
 
@@ -130,12 +106,7 @@ public class AvroPlugin implements Plugin<Project> {
         task.source(protoTask.getOutputDir());
         task.source(protoTask.getOutputs());
         task.include("**/*." + SCHEMA_EXTENSION, "**/*." + PROTOCOL_EXTENSION);
-        task.getConventionMapping().map("outputDir", new Callable<File>() {
-            @Override
-            public File call() throws Exception {
-                return getGeneratedOutputDir(project, sourceSet, JAVA_EXTENSION);
-            }
-        });
+        configurePropertyConvention(task.getOutputDir(), getGeneratedOutputDir(project, sourceSet, JAVA_EXTENSION));
 
         sourceSet.getJava().srcDir(task.getOutputDir());
 
@@ -144,49 +115,37 @@ public class AvroPlugin implements Plugin<Project> {
         compileJavaTask.source(task.getOutputs());
 
         final AvroExtension avroExtension = project.getExtensions().findByType(AvroExtension.class);
-        ConventionMapping taskMapping = conventionMapping(task);
-        taskMapping.map(OPTION_OUTPUT_CHARACTER_ENCODING, new Callable<String>() {
-            @Override
-            public String call() throws Exception {
-                String compilationEncoding = compileJavaTask.getOptions().getEncoding();
-                String extensionEncoding = avroExtension.getOutputCharacterEncoding();
-                return compilationEncoding != null ? compilationEncoding : extensionEncoding;
-            }
-        });
+
+        configurePropertyConvention(task.getOutputCharacterEncoding(), project.provider(() -> {
+            String compilationEncoding = compileJavaTask.getOptions().getEncoding();
+            String extensionEncoding = avroExtension.getOutputCharacterEncoding().getOrNull();
+            return compilationEncoding != null ? compilationEncoding : extensionEncoding;
+        }));
         return task;
     }
 
     private static void configureTaskDependencies(final Project project, final SourceSet sourceSet, final GenerateAvroJavaTask javaTask) {
-        project.getPluginManager().withPlugin("org.jetbrains.kotlin.jvm", new Action<AppliedPlugin>() {
-            @Override
-            public void execute(AppliedPlugin appliedPlugin) {
-                project.getTasks().matching(new Spec<Task>() {
-                    @Override
-                    public boolean isSatisfiedBy(Task task) {
-                        String compilationTaskName = sourceSet.getCompileTaskName("kotlin");
-                        return compilationTaskName.equals(task.getName());
+        project.getPluginManager().withPlugin("org.jetbrains.kotlin.jvm", appliedPlugin ->
+            project.getTasks().matching(task -> {
+                String compilationTaskName = sourceSet.getCompileTaskName("kotlin");
+                return compilationTaskName.equals(task.getName());
+            })
+                .all(task -> {
+                    if (task instanceof SourceTask) {
+                        ((SourceTask) task).source(javaTask.getOutputs());
+                    } else {
+                        task.dependsOn(javaTask);
                     }
-                })
-                    .all(new Action<Task>() {
-                        @Override
-                        public void execute(Task task) {
-                            if (task instanceof SourceTask) {
-                                ((SourceTask) task).source(javaTask.getOutputs());
-                            } else {
-                                task.dependsOn(javaTask);
-                            }
-                        }
-                    });
-            }
-        });
+                }));
     }
 
     private static File getAvroSourceDir(Project project, SourceSet sourceSet) {
         return project.file(String.format("src/%s/avro", sourceSet.getName()));
     }
 
-    private static File getGeneratedOutputDir(Project project, SourceSet sourceSet, String extension) {
-        return new File(project.getBuildDir(), String.format("generated-%s-avro-%s", sourceSet.getName(), extension));
+    private static Provider<Directory> getGeneratedOutputDir(Project project, SourceSet sourceSet, String extension) {
+        String generatedOutputDirName = String.format("generated-%s-avro-%s", sourceSet.getName(), extension);
+        return project.getLayout().getBuildDirectory().dir(generatedOutputDirName);
     }
 
     private static JavaCompile getCompileJavaTask(Project project, SourceSet sourceSet) {
@@ -203,14 +162,6 @@ public class AvroPlugin implements Plugin<Project> {
 
     private static SourceSet getTestSourceSet(Project project) {
         return getSourceSets(project).getByName(SourceSet.TEST_SOURCE_SET_NAME);
-    }
-
-    private static ConventionMapping conventionMapping(Object conventionAware) {
-        // TODO: try other alternatives to convention mapping
-        // Convention mapping is an internal API.
-        // Other options here:
-        // http://forums.gradle.org/gradle/topics/how_can_i_do_convention_mappings_from_java_without_depending_on_an_internal_api
-        return ((IConventionAware) conventionAware).getConventionMapping();
     }
 
     private static class NonGeneratedDirectoryFileFilter implements FileFilter {
