@@ -23,7 +23,7 @@ class FileState implements Comparable<FileState> {
     private final File file;
     private final String path;
     private String errorMessage;
-    private Set<String> duplicateTypeNames = new TreeSet<String>();
+    private Set<String> duplicateTypeNames = new TreeSet<>();
 
     FileState(File file, String path) {
         this.file = file;
